llm_agent/requirements.txt
locales
anyio==4.2.0
certifi
charset-normalizer==3.3.2
distlib==0.3.8
fastapi==0.108.0
fonttools==4.47.0
gevent==23.9.1
gevent-websocket==0.10.1
gradio_client==0.8.0
greenlet==3.0.2
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
matplotlib==3.8.2
numpy==1.26.2
oauthlib==3.2.0
openai==1.5.0
pandas==2.1.4
Pillow==10.1.0
python-socketio==5.10.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
simple-websocket==1.0.0
tokenizers==0.15.0
tornado==6.4
tqdm==4.66.1
transformers==4.36.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.25.0
uwsgitop==0.11
websockets==11.0.3